LZA-42: add default repository data settings
This change adds the pre-described repository settings for each
repository. Due to a limitation within Terraform with being unable to
loop around a tuple, the variable has been changed to a local with an
object for each repository under `repository_config`.

There is also an additional object that runs a foreach loop around the
repository config and inserts specific attributes if they do not
already exist. It also ensures that the visibility is set to private by
default unless another is opted into; ensuring that less secure options
have to be opted into.